split up VideoWebControls
